working animated sprites with basic pathfinder

+++ 3D_render/src/Player.cpp
@@ -9,94 +9,12 @@ Player::~Player()
 {
 }
 
-Player::Player(std::string const& name, int x, int z) : GameObject(name, x, z)
+Player::Player(std::string const& name, float x, float z) : GameObject(name, x, z)
 {
 }
 
 void Player::updateVelocity()
 {
-	//if (Game::event.type == SDL_KEYDOWN) {
-	//	switch (Game::event.key.keysym.sym) {
-	//	case SDLK_UP:
-	//		if (velocity.z < maxVelocity)
-	//			velocity.z += acceleration.z ;
-	//		break;
-	//	case SDLK_DOWN:
-	//		if (velocity.z > -maxVelocity)
-	//			velocity.z -= acceleration.z ;
-	//		break;
-
-	//	case SDLK_LEFT:
-	//		if (velocity.x < maxVelocity)
-	//			velocity.x += acceleration.x;
-	//		break;
-	//	case SDLK_RIGHT:
-
-	//		if (velocity.x > -maxVelocity)
-	//			velocity.x -= acceleration.x;
-	//		break;
-	//	case SDLK_z:
-	//		rotationDegrees -= 2.f;
-	//		break;
-	//	case SDLK_x:
-	//		rotationDegrees += 2.f;
-	//		break;
-	//	default:
-	//		break;
-	//	}
-	//}
-
-	//if (Game::event.type == SDL_KEYUP) {
-	//	switch (Game::event.key.keysym.sym) {
-	//	case SDLK_UP:
-	//		if (Game::almostEquals(velocity.z, 0.0f)) {
-	//			velocity.z = 0.0f;
-	//		}
-	//		if (velocity.z > 0.0f) {
-	//			velocity.z -= acceleration.z ;
-	//		}
-	//		if (velocity.z < 0.0f) {
-	//			velocity.z += acceleration.z ;
-	//		}
-	//		break;
-	//	case SDLK_DOWN:
-	//		if (Game::almostEquals(velocity.z, 0.0f)) {
-	//			velocity.z = 0.0f;
-	//		}
-	//		if (velocity.z > 0.0f) {
-	//			velocity.z -= acceleration.z;
-	//		}
-	//		if (velocity.z < 0.0f) {
-	//			velocity.z += acceleration.z;
-	//		}
-	//		break;
-	//	case SDLK_LEFT:
-	//		if (Game::almostEquals(velocity.x, 0.0f)) {
-	//			velocity.x = 0.0f;
-	//		}
-	//		if (velocity.x > 0.0f) {
-	//			velocity.x -= acceleration.x;
-	//		}
-	//		if (velocity.x < 0.0f) {
-	//			velocity.x += acceleration.x;
-	//		}
-	//		break;
-	//	case SDLK_RIGHT:
-	//		//colliderBox->y = worldSpacePosition->y + 1;
-	//		if (Game::almostEquals(velocity.x, 0.0f)) {
-	//			velocity.x = 0.0f;
-	//		}
-	//		if (velocity.x > 0.0f) {
-	//			velocity.x -= acceleration.x;
-	//		}
-	//		if (velocity.x < 0.0f) {
-	//			velocity.x += acceleration.x;
-	//		}
-	//		break;
-	//	default:
-	//		break;
-	//	}
-	//}
 
 	if (velocity.x < maxVelocity && velocity.x > -maxVelocity) {
 		velocity.x += (Game::kb[SDL_SCANCODE_A] - Game::kb[SDL_SCANCODE_D]) * acceleration.x;
